refactor: merge instructions into core
the separation of instructions from core resulted in a lot of bloat making the implementation difficult to work with. the new refactor puts instruction implementations directly on the core, and expects that a set of instructions will be passed to core on instantiation. additionally, the concept of port binding has been moved into instance methods on the core, rather than standalone functions.